changed data.py to have fixed batch size
This is because the variable batch_size was causing resource exhaustion
for GPU training over big datasets (even with a 2x GTX 1070 SLI).

Because of the way batch_generator is currently implemented, there's a
lot of useless masking that takes up memory. This still needs to be
addressed through the rewriting of AudioPrep._get_data() so that it
returns a tensor with the correct and constant batch_size and forwards
its data for correct masking inside of get_batch().